:art: 优化统一部分代码.
