+++ src/adapter/vusion/build.js
@@ -5,7 +5,7 @@ const logger = microApp.logger;
 const tryRequire = require('try-require');
 const path = require('path');
 
-module.exports = function(webpackConfig) {
+module.exports = function(wConfig) {
 
     let buildCompiler = tryRequire('vusion-cli/lib/build');
     if (!buildCompiler) {
@@ -15,38 +15,7 @@ module.exports = function(webpackConfig) {
             return null;
         }
     }
-    const finalWebpackConfig = buildCompiler.prepare(webpackConfig);
+    const webpackConfig = buildCompiler.prepare(wConfig);
 
-    const webpack = tryRequire('webpack');
-    if (!webpack) {
-        logger.error('load webpack error!');
-        return null;
-    }
-
-    // 优化处理
-    const compiler = webpack(finalWebpackConfig);
-
-    return { compiler, webpackConfig: finalWebpackConfig };
-
-    // return new Promise((resolve, reject) => {
-    //     const spinner = ora('Building for production...');
-    //     spinner.start();
-    //     compiler.run((err, stats) => {
-    //         spinner.stop();
-    //         if (err) {
-    //             // 在这里处理错误
-    //             return reject(err);
-    //         }
-
-    //         process.stdout.write(stats.toString({
-    //             colors: true,
-    //             modules: false,
-    //             children: false,
-    //             chunks: false,
-    //             chunkModules: false,
-    //         }) + '\n');
-    //         // 处理完成
-    //         resolve();
-    //     });
-    // });
+    return { webpackConfig };
 };
