--- src/adapter/webpack/index.js
@@ -1,8 +1,0 @@
-'use strict';
-
-module.exports = function(webpackConfig, isDev) {
-    if (isDev) {
-        return require('./devHot')(webpackConfig);
-    }
-    return require('./build')(webpackConfig);
-};
